refactor: extract reth mappings to module
